Implement Iris gamification module with student and teacher point calculation systems

+++ core/src/main/java/com/uniface/entity/Student.kt
@@ -24,6 +24,6 @@ class Student(
     @JsonBackReference
     var user: User? = null,
 
-    var irisPoints: Int = 0,
+    var irisPoints: Double = 0.0,
     var irisLevel: Int = 1
 )
